Make image scaling idempotent finally
Moved all the image scaling logic into a bash script so that we can
test the current dimensions of images and only scale them if they're
too big.

This way, we don't "touch" all the images that don't need scaling and
cause Git to think that all the images changed every single time.